pyproject.toml
[tool.poetry]
name = "charliepi-web"
version = "0.1.0"
description = "Web dashboard for Charlie Pi"
authors = ["Cai Williamson"]
readme = "README.md"
packages = [{include = "app.py"}]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.0"
smbus = "^1.1.post2"
rpi-gpio = "^0.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# [tool.poetry.scripts]
# serve = "app:main"
